groups: add groups schema
Thusfar groups have been represented using hydra's "warden" set of
functionality.

Unfortunately, hydra has begin splitting out the warden functionality
and it seems to be safer to punt on that integration for the immediate
future.

This is the first commti in a series to moving group management into
komadori entirely.

This will reduce the scope of hydra's responsibilities, but that seems
perfectly reasonable; hydra will still deal with the mess that is oauth2
/ oidc, and group management can't be *that* hard, right :)